+++ src/pages/Follow/components/Follower.module.scss
@@ -32,7 +32,7 @@
       overflow-x: hidden;
     }
 
-    &--font {
+    &--nickname {
       font-size: 30px;
       border: none;
       background-color: #ffffff;
