+++ src/pages/Follow/components/Follower.tsx
@@ -1,59 +1,13 @@
 import defaultImage from 'assets/images/follow/default-image.png';
 import cn from 'utils/ts/classNames';
-import {
-  acceptFollow, cancelFollow, deleteFollow, requestFollow,
-} from 'api/follow';
-import { useMutation, useQueryClient } from 'react-query';
 import { ReactComponent as Unfollow } from 'assets/svg/follow/user-remove.svg';
 import { ReactComponent as Follow } from 'assets/svg/follow/user-add.svg';
 import { useNavigate } from 'react-router-dom';
 import style from './Follower.module.scss';
-
-// 팔로우 요청 후 유저 목록을 다시 받아와 요청중 상태로 변경
-export const useRequestAndUpdate = () => {
-  const queryClient = useQueryClient();
-  const { mutate: request } = useMutation('request', (account: string) => requestFollow({ userAccount: account }), {
-    onSuccess: () => {
-      queryClient.invalidateQueries('sended');
-      queryClient.invalidateQueries('search');
-    },
-  });
-  return request;
-};
-
-// 팔로우 승인 후 받은 요청 목록을 다시 받아와 갱신
-const useAcceptFollow = () => {
-  const queryClient = useQueryClient();
-  const { mutate: accept } = useMutation('accept', (id: number) => acceptFollow({ id }), {
-    onSuccess: () => {
-      queryClient.invalidateQueries('received');
-      queryClient.invalidateQueries('follower');
-    },
-  });
-  return accept;
-};
-export const useDeleteFollow = () => {
-  const queryClient = useQueryClient();
-  const { mutate: del, data } = useMutation('delete', (account: string) => deleteFollow({ userAccount: account }), {
-    onSuccess: () => {
-      queryClient.invalidateQueries('follower');
-    },
-  });
-
-  return { del, data };
-};
-
-const useCancelFollow = () => {
-  const queryClient = useQueryClient();
-  const { mutate: cancel } = useMutation('cancel', (id: number) => cancelFollow({ id }), {
-    onSuccess: () => {
-      queryClient.invalidateQueries('sended');
-      queryClient.invalidateQueries('search');
-    },
-  });
-
-  return cancel;
-};
+import useRequestAndUpdate from '../hooks/useRequestAndUpdate';
+import useAcceptFollow from '../hooks/useAcceptFollow';
+import useDeleteFollow from '../hooks/useDeleteFollow';
+import useCancelFollow from '../hooks/useCancelFollow';
 
 interface Props {
   account: string,
@@ -77,7 +31,7 @@ export default function Follower({
       <div className={style.follower__content}>
         <button
           type="button"
-          className={cn({ [style['follower__content--font']]: true })}
+          className={cn({ [style['follower__content--nickname']]: true })}
           onClick={() => navigate(`${id}`, {
             state: {
               followId: id,
