+++ src/pages/Follow/hooks/useDeleteFollow.ts
@@ -1,0 +1,15 @@
+import { deleteFollow } from 'api/follow';
+import { useMutation, useQueryClient } from 'react-query';
+
+const useDeleteFollow = () => {
+  const queryClient = useQueryClient();
+  const { mutate: del, data } = useMutation('delete', (account: string) => deleteFollow({ userAccount: account }), {
+    onSuccess: () => {
+      queryClient.invalidateQueries('follower');
+    },
+  });
+
+  return { del, data };
+};
+
+export default useDeleteFollow;
